chore: 역할별 컴포넌트 분리
