+++ client/src/components/evaluation/BookReviewField.tsx
@@ -1,0 +1,28 @@
+import { css } from '@emotion/react';
+import { useFormContext } from 'react-hook-form';
+import RHFTextarea from '../textarea/RHFTextarea';
+
+const BookReviewField = () => {
+  const { watch } = useFormContext();
+  const rating = Number(watch('rating'));
+  const shouldWriteReview = rating <= 1 || rating >= 5;
+
+  return (
+    <div
+      css={css`
+        display: flex;
+        flex-basis: 70%;
+        flex-direction: column;
+        gap: 8px;
+      `}
+    >
+      <h2>독후감</h2>
+      <RHFTextarea
+        name="review"
+        disabled={shouldWriteReview === false}
+      />
+    </div>
+  );
+};
+
+export default BookReviewField;
